+++ test/resources/checkers/rust-multiple-error.rs
@@ -1,0 +1,8 @@
+fn foo(bar: u8) -> u8 {
+    bar + 1
+}
+
+fn spam() {
+    let i: i8 = -1;
+    foo(i);
+}